Move all of the generic functionality from the prosumer camera into the base generator
